pyproject.toml
[tool.poetry]
name = "multi-agent-block-design"
version = "0.1.0"
description = ""
authors = ["pqnhoang <hoang.pqn2123@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
genesis-world = "^0.2.1"
pillow = ">=9.2.0,<11.0"
numpy = "^1.24.0"
scipy = "^1.11.0"
matplotlib = "^3.7.0"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
pydantic = "^2.0.0"
opencv-python = "^4.8.0"
requests = "^2.31.0"
ipython = "^8.0.0"
torch = "^2.6.0"
torchvision = "^0.21.0"
pybullet = "^3.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
